# Optimize getbbox() and getextrema() routines