@homm 2024-08-14
@radarhere Ok, added, but it's just some random image from stocks)